+++ tests/conftest.py
@@ -2,6 +2,7 @@ import pytest
 from scim2_models import Group
 from scim2_models import PatchOp
 from scim2_models import PatchOperation
+from scim2_models import ScimProvider
 from scim2_models import User
 
 from scim2_client.engines.httpx2 import Client
@@ -13,9 +14,8 @@ def sync_client(httpserver):
     with Client(base_url=f"http://localhost:{httpserver.port}") as client:
         scim_client = SyncSCIMClient(
             client,
-            resource_models=[User, Group],
+            provider=ScimProvider(models=[User, Group]),
         )
-        scim_client.register_naive_resource_types()
         yield scim_client